Change float to double

+++ GaussMethod/Matrix.h
@@ -7,13 +7,13 @@ class Matrix
 public:
 	Matrix(std::string& fileName);
 
-	std::vector<std::vector<float>> matrix;
+	std::vector<std::vector<double>> matrix;
 	
 
 	void print() const;
 
 private:
-	std::vector<float> complementaryMatrix;
+	std::vector<double> complementaryMatrix;
 
 	void readMatrix(std::string& fileName);
 };
